# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job-Search Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/companies/show": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "show specific company",
                "description": "show specific company",
                "operationId": "show specific company",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Enter company id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/companies/edit": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update company",
                "description": "Edit company",
                "operationId": "update company",
                "requestBody": {
                    "description": "Pass company credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "company_id": {
                                        "type": "integer",
                                        "format": "company_id",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "comapny"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "comapny@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company update successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/companies/details": {
            "get": {
                "tags": [
                    "Company Details"
                ],
                "summary": "show specific company details",
                "description": "show specific company details",
                "operationId": "show specific company details",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Enter company id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/companies/details/create-or-update": {
            "post": {
                "tags": [
                    "Company Details"
                ],
                "summary": "Create Or Update your company details",
                "description": "Add Or Edit your company details",
                "operationId": "create or update company detail",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#components/schemas/item"
                                    },
                                    {
                                        "required": [
                                            "site",
                                            "size",
                                            "job_numbers",
                                            "description"
                                        ],
                                        "properties": {
                                            "site": {
                                                "type": "string",
                                                "format": "site",
                                                "example": "new_site"
                                            },
                                            "size": {
                                                "type": "integer",
                                                "format": "size",
                                                "example": "11"
                                            },
                                            "job_numbers": {
                                                "type": "integer",
                                                "format": "job_numbers",
                                                "example": "3"
                                            },
                                            "description": {
                                                "type": "string",
                                                "format": "description",
                                                "example": "comp_description"
                                            },
                                            "logo": {
                                                "description": "logo",
                                                "type": "string",
                                                "format": "binary"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "detail created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}